requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backports.asyncio.runner==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
click-default-group==1.2.4
condense-json==0.1.3
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
llm==0.27.1
MarkupSafe==3.0.3
matplotlib==3.9.4
mlx-metal==0.29.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
ollama==0.6.1
openai==2.16.0
optimum==2.1.0
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
psutil==7.2.2
puremagic==1.30
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ulid==3.1.0
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rpds-py==0.27.1
safetensors==0.7.0
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.8.0
tqdm==4.67.2
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
